+++ src/background/hook.rs
@@ -169,7 +169,10 @@ impl HookManager {
         if event == WinEvent::ObjectDestroy {
             return log::debug!("{:?}({:0x})", event_value, origin.address());
         }
-        log::debug!("{event_value:?} | {origin:?}");
+
+        if origin.class() == "Windows.UI.Core.CoreWindow" {
+            log::debug!("{event_value:?} | {origin:?}");
+        }
     }
 
     fn process_event(event: WinEvent, origin: Window) {